Fixed modeline creation race condition.
when the clim frame top level function is called in a separate thread,
it takes some time for the sheet to be initialized. If another function
tries to draw to the sheet, even if it has the modeline mutex, then it
will throw an error, since the sheet is NIL.

this is fixed by grabbing the mutex while creating the modeline frame,
and only releasing it once the sheet is no longer NIL.

I also added some convenience functions for reseting the modeline, and
played with the formatting of the modeline.