prototyped background colors for format items.
There wasn't a super simple way to achieve background colors in table
cells. I ended up having to create rectangles whose width and height
match the dimensions of the cell. This works for a limited set of use
cases.

This method relies on the width of the text being the same width of the
cell. If it is desireable to add spacing outside of the text width (ex
through a bounding box), the background rectangle has no way of knowing
how large it needs to be.

currently the background colors of each cell are used to display marked
windows and the current window.