Prototyped cell formatting macro
with-default-style formats the stream output to provide a uniform look
across the modeline.

I don't really like the way it works right now, there is a lot of
duplicated code which is probably unnecessary.

This macro also fixes the background size problems. The background is
now sized to fit the content, whether it is textual or graphical.