Refined default formatting macro.
I split the macro up into three macros for ease of use and modularity.
There is now a with-background macro which sets the background of any
given output. When used in conjunction with a cell (and a placeholder
rectangle due to a bug in mcclim) the entire cell background can recieve
a background color.

I also added some documentation to the functions/macros I have written
so far.