Added support for flex-box style spacers.
spacers take up all free space on the modeline. When multiple spacers
are used, the width of an individual spacer is determined by the ratio
between its weigth and the sum of all spacer weights.

currently, there is no support for a spacer configuration which would
allow elements to be centered on the modeline or to be anchored in a
specific location.