created more macros for easy presentations.
there is a lot of boilerplate in order to get a presentation translator
to do a simple action when you click on its presentation. I have boiled
these operations down to a couple macros.

the make-formatting-item macro will be renamed in the next commit to
define-formatting-item. It may need some more work, but it is the last
step (for now) in the mode line formatting process.