added crude brightness control, work on macros
adding brightness control immediately exposed issues with the
define-presentation-command and with-default-style macros. Essentially,
if you want to set the associated object of a modeline item to be
something like a string, there was no way of specifying which command
would be called when that item is clicked. This change allows you to
specify both the item and the presentation type.