created formatting-item class to track refreshing.
not all items need to be refreshed at every instance. When windows are
switched, or there is a stumpwm command, we don't need to refresh the
media player, the network manager, or even the clock. These
aforementioned items also happen to be some of the most time-consuming to
update during a refresh. by isolating the updates, the modeline can be
made to feel more 'snappy' and responsive to input.