added Bluetooth module.
The Bluetooth module displays the name of a currently connected device, and its
battery level if applicable. There are currently no interactive
capabilities within this new module. In the future, this module will
allow most everyday Bluetooth tasks to be conducted without having to
use a separate widget or go to the command line.