fixed another race condition with reset-modeline
did a bunch of on the initialization code. The lock on the application
frame thread was being released too early. I found that it needed to be
released once the frame state was set to :ENABLED rather than when the
sheet was ready to be drawn on.

I also reorganized/cleaned up the code in (redisp). It had a bunch of
superfluous operations.

I removed a bunch of code that isn't being used for list-formatters. I
will add this back in a future commit on this branch before I merge.
previous display methods should still be supported.